pyproject.toml
[tool.poetry]
name = "travel-scraper"
version = "0.1.0"
description = "Travel scraper from web"
authors = ["Ismail Gadzhiev <ismailgadzhievff@gmail.com>"]
packages = [{include = "travel_scraper"}]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.8.2"
PyYAML = "^6.0"
requests = "^2.28.2"
bs4 = "^0.0.1"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run_selenium = "travel_scraper.scraper:run_selenium"
